Fix race condition in ddahs/assignments selector
If Ddahs or Assignments were loaded before the data they depend on is
loaded, they may be improperly constructored (with empty objects
standing in for the referenced assignment/position/applicant). This
update ensure that improperly constructed data is filterd out.
