+++ frontend/src/views/instructor/routes/index.tsx
@@ -3,6 +3,7 @@ import { Redirect, Route, Switch } from "react-router-dom";
 import { InstructorPositionsView } from "../positions";
 import { InstructorAssignmentsView } from "../assignments";
 import { Landing as InstructorLanding } from "../landing";
+import { InstructorDdahsView } from "../ddahs";
 
 export function InstructorRoutes() {
     return (
@@ -20,7 +21,7 @@ export function InstructorRoutes() {
                 <InstructorAssignmentsView />
             </Route>
             <Route exact path="/tapp/ddahs">
-                <div>Not implemented yet</div>
+                <InstructorDdahsView />
             </Route>
         </Switch>
     );
